Refactor access token handling and improve error management; add IBAN generation utility and update DTOs for account and provision

+++ BankingBlazorSSR/Api/Dtos/AccountDto.cs
@@ -6,6 +6,6 @@
 public record AccountDto(
    Guid Id,
    string Iban,
-   double Balance,
+   decimal Balance,
    Guid OwnerId
 );
